add xgboost tests

+++ src/main/java/com/github/bhlangonijr/chesslib/ml/DataSet.kt
@@ -5,9 +5,9 @@ import kotlin.math.pow
 
 class DataSet(val samples: List<FeatureSet>, val featureNames: List<String>) {
 
-    fun mean(featureId: String, predicate: (FeatureSet) -> Boolean): Double {
+    fun mean(featureId: String, predicate: (FeatureSet) -> Boolean): Float {
 
-        var sum = 0.0
+        var sum = 0.0f
         var count = 0
         samples.filter { predicate.invoke(it) }
                 .filter { it.exist(featureId) }
@@ -18,14 +18,14 @@ class DataSet(val samples: List<FeatureSet>, val featureNames: List<String>) {
         return sum / max(1, count)
     }
 
-    fun variance(featureId: String, mean: Double, predicate: (FeatureSet) -> Boolean): Double {
+    fun variance(featureId: String, mean: Float, predicate: (FeatureSet) -> Boolean): Float {
 
-        var sum = 0.0
+        var sum = 0.0f
         var count = 0
         samples.filter { predicate.invoke(it) }
                 .filter { it.exist(featureId) }
                 .forEach {
-                    sum += (it.get(featureId) - mean).pow(2.0)
+                    sum += (it.get(featureId) - mean).pow(2.0f)
                     count++
                 }
         return sum / max(1, (count - 1))
@@ -36,16 +36,16 @@ class DataSet(val samples: List<FeatureSet>, val featureNames: List<String>) {
     }
 }
 
-class FeatureSet(val id: Int, val features: MutableList<Double>, val featureNameMap: Map<String, Int>) {
+class FeatureSet(val id: Int, val features: MutableList<Float>, val featureNameMap: Map<String, Int>) {
 
     fun exist(featureId: String): Boolean = featureNameMap[featureId] != null
 
-    fun get(featureId: String): Double {
+    fun get(featureId: String): Float {
         val idx = featureNameMap[featureId]
         return features[idx!!]
     }
 
-    fun getClass(): Double {
+    fun getClass(): Float {
         return features[0]
     }
